+++ src/osha/oira/client/browser/webhelpers.py
@@ -12,6 +12,7 @@ log = getLogger(__name__)
 
 class OSHAWebHelpers(WebHelpers):
 
+    hide_organisation_tab = True
     show_completion_percentage = True
 
     @memoize