+++ src/osha/oira/client/tests/test_client.py
@@ -1,13 +1,12 @@
-from osha.oira.tests.base import OiRAFunctionalTestCase
+from osha.oira.testing import OiRAFunctionalTestCase
 
 
 class ClientTest(OiRAFunctionalTestCase):
     def testTraverser(self):
-        """Test the ClientPublishTraverser in osha.oira/client.py"""
+        """Test the ClientPublishTraverser in osha.oira/client.py."""
         from ..interfaces import IOSHAClientSkinLayer
         from euphorie.client.utils import locals
-        from Products.Five.testbrowser import Browser
 
-        browser = Browser()
+        browser = self.get_browser()
         browser.open(self.portal.client.absolute_url())
         self.assertEquals(IOSHAClientSkinLayer.providedBy(locals.request), True)