+++ docs/changes.rst
@@ -7,6 +7,7 @@ Changelog
 - Mailing lists json view.
   `#739 <https://github.com/syslabcom/scrum/issues/739>`_.
   [reinhardt]
+- Support rich text in measure description (action).
 
 
 8.2.2 (2022-09-27)
@@ -27,7 +28,6 @@ Changelog
 - Change the way certificates are created #156
   [ale-rt]
 - Statistics: Incremental updates
-- Support rich text in measure description (action).
 
 
 8.1.9 (2022-09-15)