+++ docs/changes.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - Nothing changed yet.
+- Support rich text in measure description (action).
 
 
 8.2.2 (2022-09-27)
@@ -25,7 +26,6 @@ Changelog
 - Change the way certificates are created #156
   [ale-rt]
 - Statistics: Incremental updates
-- Support rich text in measure description (action).
 
 
 8.1.9 (2022-09-15)